Combining files: start with first file instead of empty PDF for metadata
If we combine multiple files, and the first file is a PDF, try
to use the first file as accumulator of other files to preserve
the metadata (author, title, moddate, producer etc) of the first
pdf, which in most cases is the main submissions.

If we cannot load the first file as pdf, fall back to starting from
empty PDF file.