Rework watermarking
Aim of the rework is to allow reproducible builds of PDFs
which include watermarks.

At the moment there are two places that change when the original
files was built with SOURCE_DATE_EPOCH and FORCE_SOURCE_DATE set:
* an object created by pikepdf for the overlay
* the /ID entry in the trailer

The new method fixes the first of the two by not using overlays,
but adding text and links directly to the first page of the document.

What remains is the /ID entry in the trailer, but all libraries tested
update the trailer in conformance to the PDF spec.
But since the trailer is always unencrypted and uncompressed, bit-wise
comparison can be done after removing the trailer /ID before the
comparison using usual tools (eg sed).

Changes made:
- separate watermark text and link target
- add extra API parameter for watermark link
- change ConverterDriver to take optional wm_link argument
- rework actual watermark code to:
  . use pymupdf
  . add text and link directly to first page instead of adding overlays