Ensure that we pass only single lines to kpse_find_files
In some instances, the file name argument of in include command
like \includegraphics can stretch multiple lines. One example
is a redefinition of \includegraphics:
	\def\includegraphics{some text with newlines}
We don't detect those as redefinitions, so try to search for
	"some text with newlines"
which is printed to stdout to be passed on to the lua script.

Make sure we escape those characters so that everything remains on
a single line.

This is in line with the arXiv requirement that filenames don't contain
strange characters (not even space, although that is not a problem),
but definitely does not contain newlines (which is from the standard
a valid part of a filename!)