ZZRM: do *not* default to set compiler to pdflatex
This was originally done before the introduction of preflight
to default to pdflatex as compiler.
But since we are nowadays using zzrm and update it from preflight,
the default "pdflatex" compiler would never be overwritten, and
so all submissions without a ZZRM would be pdflatex, even if
preflight correctly determines that we need to use "latex_dvips_ps2pdf".

Changes in this PR:
- drop the default value for compiler
- fix a missing argument to the initializer
- adjust unit tests to have empty compilers for the tests