[SSF-97] allow for quoted filenames
Most LaTeX commands that include files allow for quoted filenames,
but only double quotes seem to be supported (according to local tests).

Treat parsed filenames not only with a strip() to remove outer
whitespace, but after that strip quotes using strip('"').

This leaves a slight chance that someone uses
	\includegraphics{"foo.jpg}
though I am not sure I want to deal with that.