Try loading LaTeX log file in latin1 encoding.
The encoding of the latex log files depends on the select font
encoding, it will be T1 encoding most of the time. T1 is very close
to latin1, but not exactly, but it is the best we can do without
actually checking the output encoding in the tex code.

If even latin1 fails, read it as bytes. See
https://tex.stackexchange.com/questions/131238/what-controls-the-encoding-of-the-latex-log-file-and-how-to-change-it
for details.