ZZRM fix bug where toplevel files got lost
In case we have multiple statements for the same file,
all the previous ones got lost. e.g.
	foo.tex	toplevelfile
	foo.tex landscape
then we would only record the landscape but not the toplevelfile
information.

This is actually a "useless" specification, since landscape only applies
to .dvi files, but it can be done and is syntactically correct.

Make sure we update the same structure.