Add -halt-on-errors command line argument
We are running -interaction=batchmode, which does NOT stop
when a missing *image* file is found.
This is not how autotex did work, and also not how we want it to
work.
Adding -halt-on-error imitates the behavior of autotex which did
	latex </dev/null
and thus when an interaction was requested (any error), the progam
would terminate.

Same time: refactor command line args handling to not duplicate
the arguments in multiple places.