Fix first run failure on missing index/glossary
We added -halt-on-error which wasn't a good idea, because
the first run fails, does not generate a PDF, and the whole process
of generation fails.

We need to run without halt-on-error to generate PDFs, but check
the return code of the run.

Other option would be to check multiple needles (which was made before)
but this is fragile and depends on the reporting of LaTeX.

The return code should indicate if there was a failure that normally
would halt on error.