Fix error detection logic
Up to now we only terminated runs when there were missing files.
We tried with adding `-halt-on-error` unconditionally, but that
didn't help since some errors *on the first run* might happen.

Change the logic:
- on first run, only fail when no PDF/DVI is generated
- on subsequent runs, fail when no PDF/DVI is generated
  OR (new) when the exit code is not 0
- in addition, make sure that we remove the artifact (PDF/DVI) file
  so that surrounding converter_driver detects the error

Comments to the changes in code:
- add lots of debugging
- move the artifact removal code from "sub check_missing" to the
  iteration loop in _run_base_engine_necessary_times.
  Reason: we cannot change check_missing and make it check for error
  return code, because then the first run will fail again
- remove check_missing and depending log_inspection code

Add unit test for broken TeX file (must fail, despite producing a pdf)