Don't error in presence of nohyperref
We will not implement automatic hyperref injection, so the
nohyperref keyword is not used anymore.

When converting older corpus, we need to allow nohyperref in ZZRM,
and thus we need in principle also be able to use it in new submission,
altough there is no effect of using it with genpdf.