Add identifier parameter to API call for conversion
The default is to use a dummy timestamp provider that guarantees
that the built-in conversion driver is used.

Follow up commits will add a selection mechanism to switch the
conversion driver from local to remote based on the identifier.
For this, the identifier must be converted to a child class of
TimestampProvider, which allows computing a timestamp from the
given identifier.