Symmetric bbl version support
We currently only support bbl version 3.2 and 3.3 in TL2023
by adding an additional tree and dynamically set TEXMFAUXTREES
to support bbl version 3.3 on TL2023.

This change does the symmetric for TL2024 and TL2025, by adjusting
the TEXMFAUXTREES for the bbl version 3.2 tree added in
arxiv-texlive-base images.

Add unittest for the new case.

Also switch to new TL2025 image.