[preflight] warn when bbl version == 3.2 that TL2023 is needed
The setup is more general and will be able to adjust to further
version squeezes between supported TL versions easily.

In case the previous TL version (the one we still support) allows
for this bbl file version, we add an issue with issue type
	bbl_version_needs_previous_version
The frontend can show an appropriate warning and possible auto-select
the TL version.

Update unittests for current state.