[service] ensure that dvi based converters don't ship pdf files
Assuming a submission that selects `latex` as converter (thus,
dvi-ps-pdf route), but sneaks in `\pdfoutput=1` into the source
code. Before this change, the pdf file would be detected and the
compilation, despite only being run one time, being accepted since
a pdf was created.

By removing a pdf file if present in the two dvi based converters,
the final check for pdf files will fail, and return a failed outcome,
which is the correct one.

Another option would be to compile with
	(pdf)latex '\newcount\pdfoutput\input <MAIN>'
but I feel uncomfortable making this change.