[service] bib-processing: detect "faked" and empty bib use
revtex4-1 generates <MAIN>Notes.bib and bibstyle entries, and thus
tricks the detection algorithm believing it should run bibtex,
even when the bibliography is built-in.

Parse the aux file, detect the <MAIN>Notes.bib case, and only run
bibtex when:
- no .bbl file present
- at least one .bib file is in the @bibdata entry that is not a
  <MAIN>Notes.bib file
- also make sure that the <MAIN>Notes.bib only contains @CONTROL
  lines as created by revtex4-1, otherwise we might ignore a proper
  bib file that just happened to be named like this.

Add unittests for both cases (with Notes.bib and without).