preflight/zzrm: always use "pdftex" as compiler string
"pdftex" is an alias/shorthand, like "tex" for "etex+dvips_ps2pdf"
"pdfetex" is also a compiler string.

Both can be used to initialize, but let us ensure that compiler strings
generated are always the "short form" "pdftex".

This fixes failed selections in the frontend when pdftex is selected,
because the directives code generates "compiler: pdfetex", which is
syntactically and semantically correct, but the rest of the frontend
does not play nicely with it.